requirements.txt
python-dotenv
flask
langchain
google-generativeai
requests
streamlit
beautifulsoup4
langchain-google-genai
langchain-community